+++ src/compiler/parse_global.c
@@ -162,54 +162,6 @@ static inline Path *parse_module_path(ParseContext *c)
 
 // --- Parse import and module
 
-/**
- *
- * module_param
- * 		: TYPE_IDENT
- *		| CONST_IDENT
- *		;
- *
- * module_params ::= '{' module_param (',' module_param)* '}'
- */
-static inline bool parse_optional_module_params(ParseContext *c, Decl **decl_ref)
-{
-	SourceLoc start = c->span;
-	if (!try_consume(c, TOKEN_LBRACE)) return true;
-	if (try_consume(c, TOKEN_RBRACE)) RETURN_PRINT_ERROR_HERE("The generic parameter list cannot be empty, it needs at least one element.");
-
-	const char **tokens = NULL;
-
-	// No params
-	while (1)
-	{
-		switch (c->tok)
-		{
-			case TOKEN_TYPE_IDENT:
-			case TOKEN_CONST_IDENT:
-				break;
-			case TOKEN_COMMA:
-				RETURN_PRINT_ERROR_HERE("Unexpected ','");
-			case TOKEN_IDENT:
-				RETURN_PRINT_ERROR_HERE("The module parameter must be a type or a constant.");
-			case TOKEN_CT_IDENT:
-			case TOKEN_CT_TYPE_IDENT:
-				RETURN_PRINT_ERROR_HERE("The module parameter cannot be a $-prefixed name.");
-			default:
-				RETURN_PRINT_ERROR_HERE("Only generic parameters are allowed here as parameters to the module.");
-		}
-		vec_add(tokens, symstr(c));
-		advance(c);
-		if (!try_consume(c, TOKEN_COMMA))
-		{
-			if (!consume(c, TOKEN_RBRACE, "Expected '}'.")) return false;
-			Decl *decl = decl_new_loc(DECL_GENERIC, "", extend_loc_with_token(&start, &c->prev_span));
-			decl->generic_decl.parameters = tokens;
-			*decl_ref = decl;
-			return true;
-		}
-	}
-}
-
 
 static int generic_id = 1;
 static inline void unify_generic_decl(CompilationUnit *unit, Decl *decl)
@@ -297,13 +249,6 @@ bool parse_module(ParseContext *c, ContractDescription *contracts)
 	}
 
 	// Is this a generic module?
-	Decl *generic_decl_old = NULL;
-	if (!parse_optional_module_params(c, &generic_decl_old))
-	{
-		if (!context_set_module(c, path)) return false;
-		recover_top_level(c);
-		return true;
-	}
 	if (!context_set_module(c, path)) return false;
 	Visibility visibility = VISIBLE_PUBLIC;
 	bool weak = false;
@@ -313,18 +258,6 @@ bool parse_module(ParseContext *c, ContractDescription *contracts)
 	ASSIGN_DECL_OR_RET(Decl *generic_decl, parse_generic_decl(c), false);
 
 	if (!parse_attributes(c, &attrs, &visibility, NULL, &is_cond, NULL, &weak)) return false;
-	if (generic_decl_old)
-	{
-		SEMA_DEPRECATED(generic_decl_old, "Module-based generics is deprecated, use `<...>` instead.");
-		if (generic_decl)
-		{
-			SEMA_NOTE(generic_decl_old, "Old generics combined with new will ignore the former.");
-		}
-		else
-		{
-			generic_decl = generic_decl_old;
-		}
-	}
 	if (!parse_attach_contracts(generic_decl, contracts)) return false;
 	if (generic_decl)
 	{
@@ -359,24 +292,6 @@ bool parse_module(ParseContext *c, ContractDescription *contracts)
 					RETURN_PRINT_ERROR_AT(false, attr, "'@export' appeared more than once.");
 				c->unit->export_by_default = true;
 				continue;
-			case ATTRIBUTE_EXTERN:
-			{
-				if (vec_size(attr->exprs) != 1)
-				{
-					RETURN_PRINT_ERROR_AT(false, attr, "Expected 1 argument to '@extern(..), not %d'.",
-					                      vec_size(attr->exprs));
-				}
-				Expr *expr = attr->exprs[0];
-				if (expr->resolve_status != RESOLVE_DONE || !expr_is_const_string(expr)) RETURN_PRINT_ERROR_AT(false, expr, "Expected a constant string.");
-				if (c->unit->module->extname)
-				{
-					RETURN_PRINT_ERROR_AT(false, attr,
-					                      "External name for the module may only be declared in one location.");
-				}
-				c->unit->module->extname = expr->const_expr.bytes.ptr;
-				SEMA_DEPRECATED(attr, "'@extern' is deprecated, use '@cname' instead.");
-				continue;
-			}
 			case ATTRIBUTE_CNAME:
 			{
 				if (vec_size(attr->exprs) != 1)
@@ -521,15 +436,6 @@ static inline TypeInfo *parse_base_type(ParseContext *c)
 		CONSUME_OR_RET(TOKEN_LBRACKET, poisoned_type_info);
 		ASSIGN_EXPR_OR_RET(type_info->unresolved_type_expr, parse_expr(c), poisoned_type_info);
 		CONSUME_OR_RET(TOKEN_RBRACKET, poisoned_type_info);
-		RANGE_EXTEND_PREV(type_info);
-		return type_info;
-	}
-	if (try_consume(c, TOKEN_CT_EVALTYPE))
-	{
-		TypeInfo *type_info = type_info_new(TYPE_INFO_EVALTYPE, make_loc(c->prev_span));
-		CONSUME_OR_RET(TOKEN_LPAREN, poisoned_type_info);
-		ASSIGN_EXPR_OR_RET(type_info->unresolved_type_expr, parse_expr(c), poisoned_type_info);
-		CONSUME_OR_RET(TOKEN_RPAREN, poisoned_type_info);
 		RANGE_EXTEND_PREV(type_info);
 		return type_info;
 	}
@@ -1674,7 +1580,6 @@ static bool parse_next_is_typed_parameter(ParseContext *c, ParameterParseKind pa
 		}
 		case TYPE_TOKENS:
 		case TOKEN_TYPE_IDENT:
-		case TOKEN_CT_EVALTYPE:
 		case TOKEN_CT_TYPEOF:
 		case TOKEN_CT_TYPEFROM:
 			return true;
@@ -1988,7 +1893,7 @@ static bool parse_element_contract(ParseContext *c, ContractDescription *contrac
 	return false;
 }
 
-INLINE void attach_deprecation_from_contract(ParseContext *c, ContractDescription *contract, Decl *decl)
+INLINE void attach_deprecation_from_contract(ParseContext *c UNUSED, ContractDescription *contract, Decl *decl)
 {
 	if (contract->deprecated) vec_add(decl->attributes, contract->deprecated);
 	contract->deprecated = NULL;
@@ -2846,22 +2751,8 @@ static inline bool parse_enum_param_list(ParseContext *c, Decl*** parameters_ref
 	// If no left parenthesis we're done.
 	if (!try_consume(c, TOKEN_LPAREN)) return true;
 
-	ArrayIndex index = -1;
-	bool has_inline = !inline_index;
 	while (!try_consume(c, TOKEN_RPAREN))
 	{
-		index++;
-		bool is_inline = try_consume(c, TOKEN_INLINE);
-		if (is_inline)
-		{
-			if (!compiler.build.old_enums)
-			{
-				RETURN_PRINT_ERROR_HERE("Inline parameters are not allowed for enums.");
-			}
-			if (has_inline) RETURN_PRINT_ERROR_HERE("An enum cannot combine an inline value and a inline parameter.");
-			if (*inline_index > -1) RETURN_PRINT_ERROR_HERE("An enum may only have one inline parameter.");
-			*inline_index = index;
-		}
 		if (!parse_enum_param_decl(c, parameters_ref)) return false;
 		Decl *last_parameter = VECLAST(*parameters_ref);
 		ASSERT(last_parameter);
@@ -3302,7 +3193,8 @@ static inline bool parse_doc_contract(ParseContext *c, Expr ***list_ref, const c
 		expr->contract_expr.comment = scratch_buffer_copy();
 		if (!docs_to_comment)
 		{
-			SEMA_DEPRECATED(expr, "Not using ':' before the description is deprecated");
+			PRINT_ERROR_AT(expr, "Expected a ':' before the description.");
+			return false;
 		}
 	}
 	else
@@ -3401,7 +3293,8 @@ static inline bool parse_contract_param(ParseContext *c, ContractParam **list_re
 		}
 		else
 		{
-			SEMA_DEPRECATED(&param, "Not using ':' before the string is deprecated.");
+			PRINT_ERROR_AT(&param, "Expected a ':' before the description.");
+			return false;
 		}
 	}
 	vec_add(*list_ref, param);